Extracted MailCollector
