+++ features/fixtures/project/src/Resources/config/services.xml
@@ -17,6 +17,7 @@
 
         <service id="FSi\FixturesBundle\Admin\News" />
         <service id="FSi\FixturesBundle\Admin\PageSettings" />
+        <service id="FSi\FixturesBundle\Listener\MailCollector" />
 
     </services>
 </container>
